Refactored runtime_final
Cleaned up the code quite a bit.
Moved some logic out of __new__ and into a function. It didn't make much sense in __new__.
Mark classes and methods as final with __final__ attribute like typing.final does in Python 3.11
Added __call__ to _Final to make it more clear if a function is marked as final when called.
Moved _forbid_overriding_finals into _Final method. This allows it to access old_init_subclass via locals rather than a stored attribute.
Fixed a bug overriding non-final methods. Store the class the method was marked as final in to check against in __init_subclass__