Fixed __init_subclass__ getting incorrect class
There was an issue with the previous implementation where the stored __init_subclass__ was bound to the base class that it was accessed from rather than the subclass that it was accessed through.
This change moves the original __init_subclass__ method to a unique attribute using a uuid from which it can be later accessed on the correct class to be correctly bound.